consolidating color tools into a single module. updating century lamp
